@@ -14,8 +14,7 @@ export const useUserActivityData = () => {
     averageTransactionsPerUser: 0
   });
 
-  // Update to use a valid TimeRange value
-  const [timeRange, setTimeRange] = useState<TimeRange>('7d');
+  const [timeRange, setTimeRange] = useState<TimeRange>('week');
   const [lastUpdated, setLastUpdated] = useState<Date>(new Date());
   const [totalUniqueUsers, setTotalUniqueUsers] = useState(0);
   const [newUserGrowthRate, setNewUserGrowthRate] = useState(0);
@@ -62,7 +61,7 @@ export const useUserActivityData = () => {
 
           setTotalUniqueUsers(totalUnique);
 
-          // Calculate new user growth
+          // Calcular growth de usuarios nuevos
           const previousPeriodStart = new Date();
           previousPeriodStart.setDate(previousPeriodStart.getDate() - days * 2);
           const previousPeriodData = formattedData.filter(
